Extracted and modularized deleItem and getListDom
